refactor: move core files to own module
