+++ src/core/platform_info.rs
